% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keras_classification.R
\name{build_the_model}
\alias{build_the_model}
\title{Build and compile keras sequential models with 2 hidden layers for classification}
\usage{
build_the_model(
  train_data,
  first_layer_units,
  second_layer_units,
  classes,
  dropout_rate_1,
  dropout_rate_2,
  dense_activation_fun,
  output_activation_fun,
  optimizer_param,
  loss_param,
  metric,
  ...
)
}
\arguments{
\item{train_data}{a table of training data}

\item{first_layer_units}{integer, number of units in the first hidden layer}

\item{second_layer_units}{integer, number of units in the second hidden layer}

\item{classes}{integer, number of classes and therefore number of units in 
the output layer}

\item{dropout_rate_1}{the rate of dropout nodes for layer 1}

\item{dropout_rate_2}{the rate of dropout nodes for layer 2}

\item{dense_activation_fun}{the activation function for the hidden layers}

\item{output_activation_fun}{the activation function for the output layer}

\item{optimizer_param}{the optimizer function}

\item{loss_param}{the loss function}

\item{metric}{which metrics to monitor}

\item{...}{further arguments}
}
\value{
a compiled keras sequential model with two hidden layers
}
\description{
This function can setup and compile sequential models for classification.
}
