% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger_classification.R
\name{calculate_classification_metrics}
\alias{calculate_classification_metrics}
\title{calculate measures for classification performance}
\usage{
calculate_classification_metrics(result_table)
}
\arguments{
\item{result_table}{the result table generated by a store_*()}
}
\value{
the result_table with additional columns
}
\description{
This function calculates a bunch of classification performance metrics to help
evaluate the quality of the classification. For more info onto the calculated
values see e.g. https://en.wikipedia.org/wiki/Evaluation_of_binary_classifiers
}
