% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keras_classification.R
\name{categoric_to_factor}
\alias{categoric_to_factor}
\title{Reverse keras::to_categorical}
\usage{
categoric_to_factor(matrix)
}
\arguments{
\item{matrix}{the binary matrix which needs to be converted}
}
\value{
An integer vector with numeric factor levels
}
\description{
This function takes a binary matrix and returns one column representing
the factor levels. That way, `keras::to_categorical` can be reversed after
the machine learning step and compared to the predictions
}
