% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data_functions.r
\name{categorize}
\alias{categorize}
\title{Categorize continuous values into classes.}
\usage{
categorize(response_data, my_breaks, class_labels)
}
\arguments{
\item{response_data}{a data frame where the columns are the continuous 
response variables}

\item{my_breaks}{the intervals for the binning, specified as e.g. 
`c(-Inf, 2, 4, 6, Inf)`}

\item{class_labels}{desired names of the factor levels, Specified as e.g. 
`c("Below2", "2to4", "4to6", "Above6")` for breaks `c(-Inf, 2, 4, 6, Inf)`
or `c("Negative", "Positive")` for breaks `c(-Inf, 0, Inf)`.
At least one value needs to be a character, the others are then coerced to
class character}
}
\value{
A data frame with factor columns containing the categorized 
  response variables
}
\description{
The number and names of classes can be freely chosen.
The elements in `class_labels` need to be one less compared to the elements 
in `my_breaks`. Metrics will be calculated for each of the classes.
}
