% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data_functions.r
\name{categorize_binary}
\alias{categorize_binary}
\title{Categorize continuous values into two classes.}
\usage{
categorize_binary(response_data, my_breaks, Positive_first)
}
\arguments{
\item{response_data}{a data frame where the columns are the continuous 
response variables}

\item{my_breaks}{the intervals norders to form to bins, specified e.g. 
as `c(-Inf, 2, Inf)`}

\item{Positive_first}{logical, shall `Positive` become the first factor level? 
Defaults to `TRUE`}
}
\value{
A data frame with factor columns containing binary response variables
}
\description{
This function places continuous values into two classes. The first class is 
labelled `Negative` and the second `Positive`. By default, `Positive` is used 
as first factor level and therefore the class for which the metrics are 
later on calculated!! E.g. `True positive` are the true positives of class 
`Positive`.
}
