% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger_classification.R
\name{classification_metrics}
\alias{classification_metrics}
\title{calculate measures for classification performance}
\usage{
classification_metrics(
  result_table,
  Number_of_samples = result_table$Number_of_samples
)
}
\arguments{
\item{result_table}{the result table generated by a store_*()}

\item{Number_of_samples}{the results table column with the number of samples
to calculate metrics. The default works for ranger classification. It has to
be overwritten by the number of validation samples for keras neural networks}
}
\value{
the result_table with additional metrics
}
\description{
This function calculates a bunch of classification performance metrics to help
evaluate the quality of the classification. For more info onto the calculated
values see e.g. [Binary Classification Metrics](https://en.wikipedia.org/wiki/Evaluation_of_binary_classifiers)
}
