% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merged_list_functions.r
\name{combine_count_target}
\alias{combine_count_target}
\title{Merge all read count tables with one column from meta data each}
\usage{
combine_count_target(count_table_list, target_table)
}
\arguments{
\item{count_table_list}{a list containing the named count tables}

\item{target_table}{a dataframe containing all columns of target variables}
}
\value{
a list of merged tables
}
\description{
This function merges the prepared input tables (consisting of community_tables)
and optionally added sample data) with the prepared response data frame.
merging all count tables with all columns from the target meta data
so each combination of a given count table with one target column is
achieved.
}
