% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_taxa_functions.r
\name{create_counttable_subsets}
\alias{create_counttable_subsets}
\title{Create subsets of phyloseq objects based on ASV/OTU counts and taxa level.}
\usage{
create_counttable_subsets(subset_list, ASV_thresholds, tax_levels = NULL, ...)
}
\arguments{
\item{subset_list}{a list of phyloseq objects}

\item{ASV_thresholds}{an integer vector specifying the input
to `filter_taxa(sum(x > ASV_threshold) >= 1)`}

\item{tax_levels}{specifying the tax levels to agglomerate in the form 
of `setNames(c("To_genus", To_family), c("Genus", "Family"))`. 
Here, "To_genus" is the corresponding taxonomic level in tax_table() and 
"Genus" is appended to the name the agglomerated data.frame in the
results list for later distinction. Check taxa level using 
`colnames(tax_table(TNT_communities))`}

\item{...}{further argument passed on to filter_subsets()}
}
\value{
A list of subsetted data frames for each combination of 
  phyloseq_subset, ASV_threshold and tax_levels (+ ASV/OTU)
}
\description{
This subsetting function allows to create multiple combinations of subsets 
from a list of phyloseq objects. The lowest taxonomic level 
(usually ASV or OTU) is always included, further taxonomic levels can be 
specified as described below.
}
