% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_machine_learning.r
\name{extract_parameters}
\alias{extract_parameters}
\title{Extract parameters from named input list.}
\usage{
extract_parameters(input_list)
}
\arguments{
\item{input_list}{the input list with named items on first level}
}
\value{
a data frame with one row per list item and a column for each parameter
}
\description{
This function extracts specific parameters from first level named list items.
The data frame headers and length are set up using `initialize_results()`.
}
