% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data_functions.r
\name{extract_response_variable}
\alias{extract_response_variable}
\title{Extract response variable columns from phyloseq sample data slot.}
\usage{
extract_response_variable(response_variables, phyloseq_object)
}
\arguments{
\item{response_variables}{a vector of sample data variable names}

\item{phyloseq_object}{the according `sample_data()` will be looked up 
to receive the response columns}
}
\value{
A data frame consisting of the specified columns from `sample_data()`
}
\description{
This function extracts specified columns from the sample data slot of a 
phyloseq object. It also replaces underscores in those column headers with 
dots, as underscores are later used for string splitting.
}
