% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_machine_learning.r
\name{initialize_results}
\alias{initialize_results}
\title{This function sets up a parameter data frame.}
\usage{
initialize_results(input_list)
}
\arguments{
\item{input_list}{the list that will be used as input for machine learning}
}
\value{
an empty data frame with headers of length of input_table_list
}
\description{
Based on the length of the input list items a data frame is prepared
with headers for all information that can be extracted from the list
items names.
}
