% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keras_prepare_functions.r
\name{inputtables_to_keras}
\alias{inputtables_to_keras}
\title{Separate independent variables from response column for keras.}
\usage{
inputtables_to_keras(final_input_tables)
}
\arguments{
\item{final_input_tables}{a list of splitted input tables. The "train_set" and 
"test_set" tables need to be located at e.g `final_input_tables[[1]][["train_set"]]`
and `final_input_tables[[1]][["test_set"]]`, respectively}
}
\value{
A list of lists of splitted table_list items
}
\description{
This functions converts the prepared list into the keras tensorflow required 
shape. This means that the independent predictor variables are separated from 
the response variable. A given dummy response variable is represented by as 
many columns as it originally had levels, because the output layer of the 
neural network consists of one node per level. This step is skipped for 
non-dummy numerical response variables, e.g. as in case of regression.
}
