% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{is.dummy}
\alias{is.dummy}
\title{Tests if a vector is a dummy column}
\usage{
is.dummy(x)
}
\arguments{
\item{x}{the vector to be tested, use `sapply(x, is.wholenumber)` for a data 
frame. Strict testing using 'vapply(x, FUN = is.dummy, FUN.VALUE = logical(1)`}
}
\value{
A logical vector returning TRUE for column(s) that only contain 0 and/or 1
}
\description{
This function identifies specifically dummy variables (only 0 and 1) 
from all other column types. 1.0 as single value returns TRUE, which is a 
matter of print formatting rather of type
}
