% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{is.wholenumber}
\alias{is.wholenumber}
\title{Tests if a numeric vector is completely consisting of integers}
\usage{
is.wholenumber(x, tolerance = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{the vector to be tested, use `sapply(x, is.wholenumber)` for a data frame}

\item{tolerance}{numerical specifying the tolerance for rounding a value.
defaults to `.Machine$double.eps^0.5`}
}
\value{
A logical vector returning TRUE for column(s) that are completely integers
}
\description{
This function can be used to distinguish e.g. dummy variables (only 0 and 1) 
from other numeric columns. Applying to non numeric columns will fail. It also
allows for negative whole numbers and decimals with no rest e.g. 3.0, as this
can be a matter of print formatting rather of type
}
