% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keras_classification.R
\name{keras_classification}
\alias{keras_classification}
\title{Run keras tensorflow classification}
\usage{
keras_classification(
  Target,
  ML_object,
  Cycle,
  Epochs,
  Batch_size,
  k_fold,
  current_k_fold,
  Early_callback,
  Delay,
  step,
  the_list,
  master_grid,
  .row,
  Layer1_units,
  Layer2_units,
  Dense_activation_function,
  Dropout_layer1,
  Dropout_layer2,
  Output_activation_function,
  Optimizer_function,
  Loss_function,
  Metric
)
}
\arguments{
\item{Target}{The respective column from the master_grid}

\item{ML_object}{The respective column from the master_grid}

\item{Cycle}{The respective column from the master_grid}

\item{Epochs}{The respective column from the master_grid}

\item{Batch_size}{The respective column from the master_grid}

\item{k_fold}{The respective column from the master_grid}

\item{current_k_fold}{The respective column from the master_grid}

\item{Early_callback}{The respective column from the master_grid}

\item{Delay}{The respective column from the master_grid}

\item{step}{character declaring `training` or `prediction`}

\item{the_list}{The input tables list}

\item{master_grid}{the data frame containing all parameter combinations}

\item{.row}{current row of master_grid}

\item{Layer1_units}{integer, number of units in the first hidden layer}

\item{Layer2_units}{integer, number of units in the second hidden layer}

\item{Dense_activation_function}{the activation function for the hidden layers}

\item{Dropout_layer1}{the rate of dropout nodes for layer 1}

\item{Dropout_layer2}{the rate of dropout nodes for layer 2}

\item{Output_activation_function}{the activation function for the output layer}

\item{Optimizer_function}{the optimizer function}

\item{Loss_function}{the loss function}

\item{Metric}{which metrics to monitor}
}
\value{
a compiled keras sequential model with two hidden layers
}
\description{
This functions calls keras tensorflow using the parameter values in each row 
of the provided master_grid, using the data of the list elements.
}
