% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keras_regression.R
\name{keras_regression}
\alias{keras_regression}
\title{Run keras tensorflow regression}
\usage{
keras_regression(
  Target,
  ML_object,
  Cycle,
  Epochs,
  Batch_size,
  k_fold,
  current_k_fold,
  Early_callback,
  Delay,
  step,
  the_list,
  master_grid,
  .row,
  ...
)
}
\arguments{
\item{Target}{factor, the response variable}

\item{ML_object}{factor or char, the name of the corresponding `the_list` item}

\item{Cycle}{integer, the current repetition}

\item{Epochs}{keras, integer, how many times should the whole data set be 
passed through the network?}

\item{Batch_size}{keras, integer, how many samples before updating the weights?}

\item{k_fold}{integer, the total number of k_folds for cross validation}

\item{current_k_fold}{integer, the current k_fold in range 1 : k_fold}

\item{Early_callback}{keras, string, a callback metric}

\item{Delay}{keras, integer, wait for how many epochs before callback happens?}

\item{step}{character declaring `training` or `prediction`}

\item{the_list}{The input tables list}

\item{master_grid}{the data frame containing all parameter combinations}

\item{.row}{current row of master_grid}

\item{...}{additional features passed by pmap call}
}
\value{
a compiled keras sequential model with two hidden layers#'
}
\description{
This functions calls keras tensorflow using the parameter values in each row 
of the provided master_grid, using the data of the list elements. Please have
a look at the keras [fit doc](https://keras.rstudio.com/reference/fit.html)
for explanation on the keras related variables, the arguments are beginning 
with "keras" in the description. Except for `the list`, `master_grid` and `.row`
all arguments need to be column names of `master_grid`
}
