% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_taxa_functions.r
\name{otu_table_to_df}
\alias{otu_table_to_df}
\title{Turn subsets of phyloseq objects into relative abundance data.frames.}
\usage{
otu_table_to_df(subset_list)
}
\arguments{
\item{subset_list}{a list of phyloseq objects}
}
\value{
A list of subsetted community tables as data.frames
}
\description{
The provided list of phyloseq objects is turned to relative abundances 
in percentage. The community tables are converted to data.frames and stored
in a list. NaNs are turned into 0.
}
