% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merged_list_functions.r
\name{oversample}
\alias{oversample}
\title{Replicate observations with added noise.}
\usage{
oversample(splitted_list, copy_number, noise_factor)
}
\arguments{
\item{splitted_list}{a list of complimentary training and test data sets named
"train_set" and "test_set", e.g `mylist[[1]][["train_set"]]` and 
`mylist[[1]][["test_set"]]` for the first list item.}

\item{copy_number}{an integer specifying the number of copies, 0 means no 
oversampling takes place and `noise_factor` is ignored}

\item{noise_factor}{a value >= 0 specifing the relative amount of noise randomly 
added or substracted to/from the original value e.g. 0.05 == +-5 \% noise}
}
\value{
A list of lists, the list item name is updated to reflect the number 
  of copies and the noise value
}
\description{
This function creates copies of existing observations in the training set 
with a specified amount of noise added to each numeric variable/observation 
combination with a non-zero value using data.table. This includes taxa abundance and also 
sample data variables. Pure integer columns are ignored, as they are either 
all 0s or dummy columns (only 0 and 1) and represent factor levels. The 
response variable is also ignored. Copies are named according to this scheme: 
original: "Sample_1", first copy: "Sample_1.1", second copy: "Sample_1.2" etc
The actual work is performed by internal `oversample_and_noise()` function.
}
