% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merged_list_functions.r
\name{oversample_and_noise}
\alias{oversample_and_noise}
\title{Replicate observations with added noise.}
\usage{
oversample_and_noise(split_table, copy_number, noise_factor)
}
\arguments{
\item{split_table}{a list consisting of "train_set" and "test_set"}

\item{copy_number}{an integer specifying the number of copies, 0 means no 
oversampling takes place and `noise_factor` is ignored}

\item{noise_factor}{a value >= 0 specifing the relative amount of noise randomly 
added or substracted to/from the original value e.g. 0.05 == +-5 \% noise}
}
\value{
A list of lists, the list item name is updated to reflect the number 
  of copies and the noise value
}
\description{
This function gets called by `oversample`, see description there.
}
