% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger_classification.R
\name{prediction_accuracy}
\alias{prediction_accuracy}
\title{Calculate accuracy percentage for ranger predicted_rfs}
\usage{
prediction_accuracy(predicted_rf, test_set)
}
\arguments{
\item{predicted_rf}{the ranger prediction object}

\item{test_set}{the test_set containing the true values}
}
\value{
The accuracy in percentage
}
\description{
This function calculates predicted_rf accuracy, which means it counts the
amount of true positives and true negatives.
}
