% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_taxa_functions.r
\name{process_subsets_multi_tax}
\alias{process_subsets_multi_tax}
\title{Create subsets of phyloseq objects with subsetting based on ASV reads and taxa level.}
\usage{
process_subsets_multi_tax(subset_list, ASV_thresholds, tax_levels)
}
\arguments{
\item{subset_list}{a list of phyloseq objects}

\item{ASV_thresholds}{an integer vector specifying the input to filter_taxa(sum(x > ASV_threshold) >= 1)}

\item{tax_levels}{specifying the tax levels to agglomerate in the form of setNames(c("To_genus", To_family), c("Genus", "Family"))}
}
\value{
A list of subsetted data frames for each combination of phyloseq_subset, ASV_threshold and tax level + ASV
}
\description{
This subsetting function allows to create multiple combinations of subsets from a
list of phyloseq object based on ASV thresholds for different taxa level, always including
the ASV level. The read counts of the new subsets are then turned into relative abundance
and the count tables are finally stored as list of data.frames. NaNs are turned into 0.
}
