% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger_regression.r
\name{ranger_regression}
\alias{ranger_regression}
\title{Run ranger regression with parameters of data.frame rows.}
\usage{
ranger_regression(
  master_grid,
  Target,
  ML_object,
  Cycle,
  Number_of_trees,
  Mtry_factor,
  .row,
  the_list,
  step,
  ...
)
}
\arguments{
\item{master_grid}{the data frame containing all parameter combinations}

\item{Target}{character, the response variable}

\item{ML_object}{factor or character, the name of the corresponding `the_list` item}

\item{Cycle}{integer, the current repetition}

\item{Number_of_trees}{ranger, integer, number of trees per forest}

\item{Mtry_factor}{ranger, factor to multiply default ranger mtry argument}

\item{.row}{current row of master_grid}

\item{the_list}{The input tables list}

\item{step}{character declaring `training` or `prediction`}

\item{...}{further parameters passed on to subfunctions}
}
\value{
a data frame with results and metrics for each row of the master_grid
}
\description{
This functions calls ranger using the parameter values in each row of the 
provided master_grid, using the data of the list elements. Please have
a look at the [ranger doc](https://cran.r-project.org/web/packages/ranger/ranger.pdf)
for explanation on the ranger related variables, the arguments are beginning 
with "ranger" in the description. Except for `the list`, `master_grid` and `.row`
all arguments need to be column names of `master_grid`
}
