% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merged_list_functions.r
\name{split_data}
\alias{split_data}
\title{Split data frames according to a ratio into training and test sets}
\usage{
split_data(merged_list, split_ratios)
}
\arguments{
\item{merged_list}{a list of data frames}

\item{split_ratios}{a numerical vector of ratios from 0 to 1 indicating the 
training fraction after the split, e.g. `c(0.75, 0.8)` for two rounds of
splits with resulting training sets of 75 and 80 percent of the total samples}
}
\value{
A named nested list, containing both parts of the splitted data frame. 
  The name is updated to reflect the split ratio
}
\description{
This function applies a split to the data frames contained in the input list. 
It first groups the data frames based on their number of rows. It then iterates
over the data frames in each group and the provided split ratios to separate
each data frame into a train and a test set. A nested list is returned having
the names as first level and the respective split data sets on second level.
}
