% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merged_list_functions.r
\name{split_dataset}
\alias{split_dataset}
\title{Split a list of data frames according to the ratio into training and test data}
\usage{
split_dataset(merged_column_list, split_ratios)
}
\arguments{
\item{merged_column_list}{a list with data frames and the merged targed variable as last column}

\item{split_ratios}{a numerical vector of ratios from 0 - 1 indicating the training fraction after the split}
}
\value{
A list of lists, each data frame is split into train and test set, the name is updated to reflect the ratio value
}
\description{
This function applies a split to the data frames contained in the input list. It first groups the
data frames based on their number of rows. The same number indicates they originate from the same main
subset, but different ASV thresholds or tax level are present. Therefore, the samples for training and test
should be the same for comparison. Such data frames are combined as nested lists in the grouped_list. This
list is then split into train and test set for all data frames and ratios provided. The output list is already
flattened and with updated name, so the names are first level and train and test data frames are second level
}
