% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_phyloseq_functions.r
\name{standardize_phyloseq_headers}
\alias{standardize_phyloseq_headers}
\title{Change naming style of the taxa in the phyloseq object.}
\usage{
standardize_phyloseq_headers(phyloseq_object, taxa_name, use_sequences)
}
\arguments{
\item{phyloseq_object}{The phyloseq object with the taxa_names() to be modified.}

\item{taxa_name}{The leading name of your taxa, e.g. `ASV` or `OTU`}

\item{use_sequences}{Logical indicating whether your current taxa_names()
are sequences which you want to have moved to the refseq() slot of the 
returning phyloseq object}
}
\value{
The modified phyloseq object.
}
\description{
This function adapts the rows and the columns of the corresponding
phyloseq object to improve the reading. It adds 0s based
on the total amount of ASVs to enable correct sorting in R. If you imported
your data from dada2 and your current taxa names are the sequences (ASVs),
you can specify `use_sequences = TRUE` to move the sequences to the phyloseq
refseq() slot before renaming takes place
}
