% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger_classification.R
\name{store_multi}
\alias{store_multi}
\title{Store results from ranger classification training and prediction}
\usage{
store_multi(
  trained_rf,
  predicted_rf = NULL,
  confusion_matrix,
  timings,
  n_classes,
  step,
  ...
)
}
\arguments{
\item{trained_rf}{the ranger object generated by training with `ranger()`}

\item{predicted_rf}{the ranger object generated by prediction with `predict()`.
defauls to `NULL`}

\item{confusion_matrix}{the confusion matrix generated from the results of
`predicted_rf`, defaults to `NULL`}

\item{timings}{the timings resulting from the `system.time()` call}

\item{n_classes}{the number of classes for classification}

\item{step}{character declaring whether `training` or `prediction`}

\item{...}{parameters passed on to `prediction_accuracy`}
}
\value{
A data frame with one row per ranger run
}
\description{
This function extracts information from the ranger objects generated by
or prediction and stores them in a data.frame. It also logs some additional
values such as used memory and elapsed time and calls the functions 
`calculate_classification_metrics` and `prediction_accuracy` for more data.
}
