% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger_regression.r
\name{store_regression}
\alias{store_regression}
\title{Store results from ranger regression training and prediction.}
\usage{
store_regression(
  trained_rf,
  predicted_rf = NULL,
  step,
  training_data,
  test_data = NULL
)
}
\arguments{
\item{trained_rf}{the ranger object generated by training with `ranger()`}

\item{predicted_rf}{the ranger object generated by prediction with `predict()`,
(default: `NULL`)}

\item{step}{character declaring whether `training` or `prediction` occurs}

\item{training_data}{data frame the ranger object was trained with}

\item{test_data}{data frame the ranger object used for prediction}
}
\value{
A data frame with one row per ranger run and class
}
\description{
This function extracts information from the ranger objects generated by
training or prediction and stores them in a data.frame. It compares the pre-
dicted values in training and prediction to their true values and calculates
various measures to describe the difference. `Rsquared_all` refers to R2 calcu-
lated on all samples instead of only out-of-bag samples in `Rsquared`. This is
only relevant for training, as there is no out-of-bag in prediction.
}
