% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taxa_vector_list}
\alias{taxa_vector_list}
\title{Dictionary for taxonomic level conversion.}
\format{A list with two vectors of taxonomic level Genus and Family}
\usage{
taxa_vector_list
}
\description{
A list with vectors linking IDs (ASV/OTU/...) to their annotated taxonomic levels.
}
\keyword{datasets}
