% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testps}
\alias{testps}
\title{Renamed TNT communities.}
\format{A phyloseq object with 105 samples, 50 taxa and 93 sample variables}
\usage{
testps
}
\description{
The phyloseq object `TNT_communities`with adjusted taxa names.
}
\keyword{datasets}
