% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TNT_communities}
\alias{TNT_communities}
\title{Microbial communities in ammunition contaminated sediments.}
\format{A phyloseq object with 105 samples, 50 taxa and 93 sample variables}
\usage{
TNT_communities
}
\description{
A dataset in form of a phyloseq object containing an ASV count table, 
  sample data and taxonomy data .
}
\keyword{datasets}
