% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_taxa_functions.r
\name{to_relative_abundance}
\alias{to_relative_abundance}
\title{Turn otu_table() of phyloseq objects to relative abundance}
\usage{
to_relative_abundance(subset_list)
}
\arguments{
\item{subset_list}{a list of phyloseq objects}
}
\value{
A list of phyloseq objects
}
\description{
The provided list of phyloseq objects is turned to relative abundances 
in percentage.
}
