% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy_level_functions.r
\name{translate_ID}
\alias{translate_ID}
\title{Translate unique IDs such as ASV/OTUs to specified tax level}
\usage{
translate_ID(ID, translate_to, lookup_table, na.rm = TRUE)
}
\arguments{
\item{ID}{a string with the ID name you want to translate}

\item{translate_to}{a string with the desired taxonomic rank for translation}

\item{lookup_table}{the list with all the lookup vectors created 
by create_taxonomy_lookup().}

\item{na.rm}{a logical, default is TRUE. All IDs that were assigned as NA 
on the specified taxonomic level will be removed. Use FALSE if same 
length as input vector is required}
}
\value{
A character vector of taxonomic annotations on the corresponding 
  taxonomic level per provided ID
}
\description{
This function translates a vector of given identifiers to the specified 
taxonomy level. The  lookup table required is generated by `create_taxonomy_lookup()`.
}
