% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data_functions.r
\name{add_sample_data}
\alias{add_sample_data}
\title{Select sample data variables and add them to the count tables.}
\usage{
add_sample_data(sample_data_names, phyloseq_object, community_tables)
}
\arguments{
\item{sample_data_names}{a vector of sample data variable names}

\item{phyloseq_object}{the phyloseq object containing the sample data}

\item{community_tables}{the list with relative abundance community tables}
}
\value{
A list of the merged community + sample data tables as data.frames
}
\description{
This function selects specified columns from the sample data of a phyloseq object
and iteratively merges each count table with the specified sample data columns.
Merging occurs using the row.names.
}
