% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keras_classification.R
\name{build_the_model}
\alias{build_the_model}
\title{Build and compile keras sequential models with 2 hidden layers for classification}
\usage{
build_the_model(
  train_data,
  Layer1_units,
  Layer2_units,
  classes,
  Dropout_layer1,
  Dropout_layer2,
  Dense_activation_function,
  Output_activation_function,
  Optimizer_function,
  Loss_function,
  Metric,
  ...
)
}
\arguments{
\item{train_data}{a table of training data}

\item{Layer1_units}{integer, number of units in the first hidden layer}

\item{Layer2_units}{integer, number of units in the second hidden layer}

\item{classes}{integer, number of classes and therefore number of units in 
the output layer}

\item{Dropout_layer1}{the rate of dropout nodes for layer 1}

\item{Dropout_layer2}{the rate of dropout nodes for layer 2}

\item{Dense_activation_function}{the activation function for the hidden layers}

\item{Output_activation_function}{the activation function for the output layer}

\item{Optimizer_function}{the optimizer function}

\item{Loss_function}{the loss function}

\item{Metric}{which metrics to monitor}

\item{...}{further arguments}
}
\value{
a compiled keras sequential model with two hidden layers
}
\description{
This function can setup and compile sequential models for classification.
}
