% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_taxa_functions.r
\name{create_community_table_subsets}
\alias{create_community_table_subsets}
\title{Create subsets of phyloseq objects based on ASV/OTU counts and taxa level.}
\usage{
create_community_table_subsets(
  subset_list,
  thresholds,
  tax_ranks = NULL,
  taxa_prefix,
  ...
)
}
\arguments{
\item{subset_list}{a list of phyloseq objects}

\item{thresholds}{an integer vector specifying the input
to `filter_taxa(sum(x > threshold) >= 1)`}

\item{tax_ranks}{specifying the tax ranks to agglomerate in the form 
of `setNames(c("To_genus", To_family), c("Genus", "Family"))`. 
Here, "To_genus" is the corresponding taxonomic level in tax_table() and 
"Genus" is appended to the name the agglomerated data.frame in the
results list for later distinction. Check taxa rank using 
`colnames(tax_table(TNT_communities))`}

\item{taxa_prefix}{The leading name of your taxa, e.g. `ASV` or `OTU`, 
must not contain an underscore or white space}

\item{...}{further argument passed on to filter_subsets()}
}
\value{
A list of subsetted community tables for each combination of 
  phyloseq_subset, thresholds and tax_ranks (+ ASV/OTU)
}
\description{
This subsetting function allows to create multiple combinations of subsets 
from a list of phyloseq objects. The lowest taxonomic level 
(usually ASV or OTU) is always included, further taxonomic levels can be 
specified as described below. The call to `phyloseq::tax_glom` can be slow, 
therefore package `speedyseq` is used, if installed.
}
