% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy_level_functions.r
\name{create_taxonomy_lookup}
\alias{create_taxonomy_lookup}
\title{Creates a dictionary list to translate taxonomic levels}
\usage{
create_taxonomy_lookup(phyloseq_object, taxonomic_levels)
}
\arguments{
\item{phyloseq_object}{the phyloseq object to extract the taxonomic information from}

\item{taxonomic_levels}{a character vector with taxa levels 
e.g. from `colnames(phyloseq::tax_table(phyloseq_object))`}
}
\value{
A list of vectors for each tax level with the corresponding unique 
  identifier as row name
}
\description{
This function creates a list containing the entries for a higher taxonomic 
level in relation to their low level unique identifier (i.e. OTUs or ASVs). 
The taxonomic levels can be specified and need to be present in `tax_table()`` 
of the respective phyloseq object
}
