% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_taxa_functions.r
\name{filter_subsets}
\alias{filter_subsets}
\title{Remove ASVs from a phyloseq object.}
\usage{
filter_subsets(phyloseq_object, threshold = 0, num_samples = 1)
}
\arguments{
\item{phyloseq_object}{the object to be filtered}

\item{threshold}{this amount of reads have to appear for an ASV, than all ASV reads are kept}

\item{num_samples}{the number of times the threshold has to be met}
}
\value{
The subsetted phyloseq object
}
\description{
This version of the filter_taxa function can be used with lists and lapply or for loops.
}
