% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keras_regression.R
\name{keras_regression}
\alias{keras_regression}
\title{Run keras tensorflow regression}
\usage{
keras_regression(
  Target,
  ML_object,
  Cycle,
  Epochs,
  Batch_size,
  k_fold,
  current_k_fold,
  Early_callback,
  Delay,
  step,
  the_list,
  master_grid,
  .row,
  ...
)
}
\arguments{
\item{Target}{The respective column from the master_grid}

\item{ML_object}{The respective column from the master_grid}

\item{Cycle}{The respective column from the master_grid}

\item{Epochs}{The respective column from the master_grid}

\item{Batch_size}{The respective column from the master_grid}

\item{k_fold}{The respective column from the master_grid}

\item{current_k_fold}{The respective column from the master_grid}

\item{Early_callback}{The respective column from the master_grid}

\item{Delay}{The respective column from the master_grid}

\item{step}{character declaring `training` or `prediction`}

\item{the_list}{The input tables list}

\item{master_grid}{the data frame containing all parameter combinations}

\item{.row}{current row of master_grid}

\item{...}{additional features passed by pmap call}
}
\value{
a compiled keras sequential model with two hidden layers
}
\description{
This functions calls keras tensorflow for regression using the parameter 
values in each row of the provided master_grid, using the data of the list elements.
}
