% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merged_list_functions.r
\name{oversample_and_noise}
\alias{oversample_and_noise}
\title{Replicate observations with added noise.}
\usage{
oversample_and_noise(split_table, copy_number, noise_factor)
}
\arguments{
\item{split_table}{a list consisting of "train_set" and "test_set"}

\item{copy_number}{an integer specifying the number of copies, 0 means no 
oversampling takes place and `noise_factor` is ignored}

\item{noise_factor}{a value between 0 and 100 (100 equals 10,000 \% noise) 
specifing the relative amount of noise randomly added or substracted to/from 
the original value. e.g. 0.05 == +-5 \% noise}
}
\value{
A list of lists, the list item name is updated to reflect the number 
  of copies and the noise value
}
\description{
This function gets called by `oversample` and creates copies of samples with
a specified amount of random noise. For more information see the docstring of
`oversample`. 
The anonymous data.table function for replication and noise addition acts on
all columns which are numeric, but neither the response variable or 0/1 only
columns, as these are either dummy variables or empty columns. It copies each
row of a train set as often as defined by `copy_number` and adds or substracts
randomly the amount of `noise_factor`
}
