% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger_regression.r
\name{ranger_regression}
\alias{ranger_regression}
\title{Run ranger regression with parameters of data.frame rows.}
\usage{
ranger_regression(
  master_grid,
  Target,
  ML_object,
  Cycle,
  Number_of_trees,
  Mtry_factor,
  .row,
  the_list,
  step,
  ...
)
}
\arguments{
\item{master_grid}{the data frame containing all parameter combinations}

\item{Target}{The respective column from the master_grid}

\item{ML_object}{The respective column from the master_grid}

\item{Cycle}{The respective column from the master_grid}

\item{Number_of_trees}{The respective column from the master_grid}

\item{Mtry_factor}{The respective column from the master_grid}

\item{.row}{current row of master_grid}

\item{the_list}{The input tables list}

\item{step}{character declaring `training` or `prediction`}

\item{...}{further parameters passed on to subfunctions}
}
\value{
a data frame with results and metrics for each row of the master_grid
}
\description{
This functions calls ranger using the parameter values in each row of the 
provided master_grid, using the data of the list elements.
}
