% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_taxa_functions.r
\name{to_relative_abundance}
\alias{to_relative_abundance}
\title{Turn subsets of phyloseq objects into relative abundance data.frames.}
\usage{
to_relative_abundance(subset_list)
}
\arguments{
\item{subset_list}{a list of phyloseq objects}
}
\value{
A list of subsetted community tables as data.frames
}
\description{
The provided list of phyloseq objects is turned to relative abundances 
in percentage. The community tables are converted to data.frames and stored
in a list. NaNs are turned into 0.
}
